Change Scala Sources Name
